Zmieniono typy argumentów

+++ chinski_sop/sop/memory/logical_memory.cpp
@@ -11,15 +11,15 @@ void sop::memory::LogicalMemory::setPageTableSize(int8_t size)
 {
   this->page_table_size=size;
 }
-int8_t sop::memory::LogicalMemory::getPageTableSize()
+uint8_t sop::memory::LogicalMemory::getPageTableSize()
 {
   return page_table_size;
 }
-sop::memory::page* sop::memory::LogicalMemory::getPage(int8_t page_nr)
+sop::memory::page* sop::memory::LogicalMemory::getPage(uint8_t page_nr)
 {
   return &page_table[page_nr];
 }
-void sop::memory::LogicalMemory::setPage(int8_t frame_nr,char val_inv,int8_t page_nr)
+void sop::memory::LogicalMemory::setPage(uint16_t frame_nr,char val_inv,int8_t page_nr)
 {
   this->page_table[page_nr].frame_number=frame_nr;
   this->page_table[page_nr].valid_invalid=val_inv;
