+++ chinski_sop/sop/memory/logical_memory.h
@@ -18,14 +18,14 @@ class LogicalMemory
 public:	
   explicit LogicalMemory(int8_t number_of_page);//konstruktor tworz¹cy tablicê stron o podanej liczbie stron
   void setPageTableSize(int8_t size);//ustawia liczbê stron
-  int8_t getPageTableSize();//zwraca liczbê stron
-  page* getPage(int8_t page_nr);//zwraca wskazan¹ stronê
-  void setPage(int8_t frame_nr,char val_inv,int8_t page_nr);//ustawia poszczególnej strony i jej parametrów
+  uint8_t getPageTableSize();//zwraca liczbê stron
+  page* getPage(uint8_t page_nr);//zwraca wskazan¹ stronê
+  void setPage(uint16_t frame_nr,char val_inv,int8_t page_nr);//ustawia poszczególnej strony i jej parametrów
   
 protected:
 
 private:
-	int8_t page_table_size; //zmienna odpowiadaj¹ca za liczbê  stron
+	uint8_t page_table_size; //zmienna odpowiadaj¹ca za liczbê  stron
   boost::shared_ptr<page[]> page_table;//tabela stron, rekordów tyle ile ramek
 };
 //page_table = boost::shared_ptr<page[]>(new page[page_table_size]); w konruktorze zrobic
